// src/app/secure/accounts/services/models/secure.user.profile.ts
export class UserProfile {
    id: number;
    userName: string;
    firstName: string;
    middleName: string;
    lastName: string;
    title: string;
    companyName: string;
    description: string;
    highestQualification: string;
    aboutMe: string;
    contact: Array<Contact>;
    address: Array<Address>;
    qualifications: Array<Qualification>;
    certifications: Array<Certification>;
    specialities: Array<Speciality>;
}
export class Speciality { }
export class Certification { }
export class Qualification { }
export class Address {

}
export class Contact {

}
export class Personal {


}